more modules

+++ download_html.py
@@ -6,13 +6,13 @@ from application import Module
 import os
 
 modules = Module.query.all()
+files_already = os.listdir(os.path.join("data", "syllabus_html"))
 
-for module in modules:
-    if module.dept_code == "PSYC":
-        print module.url
+query_departments = ["PSYC", "ANAT", "BIOC", "BIOL", "BIOS", "CELL", "NEUR", "PHAR", "PHOL", "PALS", "PHAY", "ARCL"]
+
 # TODO prevent this from asking for duplicates
 urls = [{"code": module.id, "url": module.machine_url} for module in modules if
-        module.dept_code == "PSYC" and module.machine_url is not None]
+        module.dept_code in query_departments and module.machine_url is not None]
 
 # Save all the results into files.
 
@@ -22,5 +22,6 @@ def html_output_file(module_code):
 
 
 for url in urls:
-    print url
-    urllib.urlretrieve(url["url"], html_output_file(url["code"]))
+    if url['code'] + '.html' not in files_already:
+        print url
+        urllib.urlretrieve(url["url"], html_output_file(url["code"]))
